Added Timer
Added Timer Node. Defined the Functions namespace. Moving towards playing audio sequentially from Functions.
