+++ soundtest.tscn
@@ -1,8 +1,8 @@
 [gd_scene load_steps=2 format=3 uid="uid://dvbsa3b2xhodc"]
 
-[ext_resource type="Script" path="res://PlayFunctions.cs" id="1_1erne"]
+[ext_resource type="Script" path="res://PlayFunctions.cs" id="1_67hnl"]
 
 [node name="Root" type="Node"]
 
 [node name="Sounds" type="Node" parent="."]
-script = ExtResource("1_1erne")
+script = ExtResource("1_67hnl")
